requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
datasets==2.14.5
diffusers==0.31.0
dill==0.3.7
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2023.6.0
grpcio==1.68.0
huggingface==0.0.1
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
inquirerpy==0.3.4
Jinja2==3.1.4
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.8.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
networkx==3.4.2
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pfzy==0.3.4
Pillow==10.0.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.29.0
psutil==6.1.0
pyarrow==18.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
six==1.16.0
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
torch==2.1.2
torch-tb-profiler==0.4.3
torchvision==0.16.2
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.0
zipp==3.21.0
